refactor: 전역 store에서 사용하는 type을 types/sotres 경로에서 관리 하도록 수정
