--- src/types/confirmModalType.ts
@@ -1,5 +1,0 @@
-export interface ConfirmModalType {
-  content: string
-  yesButtonText?: string
-  noButtonText?: string
-}
