+++ src/types/stores/confirmType.ts
@@ -1,0 +1,18 @@
+export interface ConfirmModalType {
+  content: string
+  yesButtonText?: string
+  noButtonText?: string
+}
+
+export interface ConfirmState {
+  content: string
+  yesButtonText: string
+  noButtonText: string
+  clickYes: (value: unknown) => void
+  clickNo: (value: unknown) => void
+  openConfirmModal: ({
+    content,
+    yesButtonText,
+    noButtonText,
+  }: ConfirmModalType) => Promise<boolean>
+}
